added image url generation for email | email working perfect
