+++ main.py
@@ -1,4 +1,5 @@
 from fastapi import FastAPI, Depends, HTTPException, UploadFile, File, BackgroundTasks
+from fastapi.staticfiles import StaticFiles
 from sqlalchemy.orm import Session
 from fastapi.middleware.cors import CORSMiddleware
 from sqlalchemy import func
@@ -9,24 +10,19 @@ from database import get_db, engine
 import models
 from pydantic import BaseModel, field_validator
 from fastapi.responses import FileResponse, JSONResponse
-import openpyxl
 from openpyxl import Workbook
 import tempfile
 import os
 import shutil
 from enum import Enum
-import pandas as pd
 import joblib
 from statementExtractor import extract_transactions
-from email.mime.multipart import MIMEMultipart
-from email.mime.text import MIMEText
-import smtplib
 from dotenv import load_dotenv
 import os
 from apscheduler.schedulers.background import BackgroundScheduler
 from apscheduler.triggers.cron import CronTrigger
 import atexit
-
+from service.mail_service import _send_monthly_report_logic, send_email, scheduled_report_job
 load_dotenv()
 
 
@@ -37,6 +33,8 @@ model = joblib.load('categoryFinder.pkl')
 models.Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
+
+app.mount("/static", StaticFiles(directory="static"), name="static")
 
 origins = [
     "http://localhost:5173",
@@ -777,79 +775,6 @@ def get_daily_expenses(month: int, year: int, db: Session = Depends(get_db)):
         ]
     }
 
-def _send_monthly_report_logic(db: Session, year: int, month: int):
-    """
-    Contains the core logic for generating the monthly report for a specific month and year.
-    Returns a dictionary with 'subject' and 'body' if a report is generated, otherwise None.
-    """
-    report_date = datetime(year, month, 1)
-    start_date = datetime(year, month, 1)
-    end_date = start_date + relativedelta(months=1)
-
-    budget = db.query(models.Budget).filter(
-        models.Budget.created_at >= start_date,
-        models.Budget.created_at < end_date
-    ).first()
-
-    template_file = 'email_templates/monthly_report_template.html' if budget else 'email_templates/monthly_report_template_no_budget.html'
-
-    expenses = db.query(models.Expense).filter(
-        models.Expense.date >= start_date,
-        models.Expense.date < end_date
-    ).all()
-
-    total_spent = sum(expense.amount for expense in expenses)
-    total_saved = (budget.monthly_income - total_spent) if budget else 0
-
-    last_month_start = start_date - relativedelta(months=1)
-    past_expenses = db.query(models.Expense).filter(
-        models.Expense.date >= last_month_start,
-        models.Expense.date < start_date
-    ).all()
-    past_total_spent = sum(expense.amount for expense in past_expenses)
-    percent_change_expenses = ((total_spent - past_total_spent) / past_total_spent * 100) if past_total_spent > 0 else "N/A"
-
-    overspent_categories = []
-    if budget and budget.category_budgets:
-        category_expenses_q = db.query(models.Expense.category_id, func.sum(models.Expense.amount).label('total'))\
-            .filter(models.Expense.date >= start_date, models.Expense.date < end_date)\
-            .group_by(models.Expense.category_id).all()
-
-        category_expenses = {item.category_id: item.total for item in category_expenses_q}
-
-        for cat_id, total_spent_for_cat in category_expenses.items():
-            budget_for_cat = budget.category_budgets.get(str(cat_id))
-            if budget_for_cat and total_spent_for_cat > budget_for_cat:
-                overspent_categories.append(CATEGORIES[cat_id])
-
-    budget_used_percent = (total_spent / budget.monthly_income * 100) if budget and budget.monthly_income > 0 else 0
-    savings_goal_reached = total_saved >= budget.saving_goal if budget else False
-
-    top_expenses = sorted(expenses, key=lambda x: x.amount, reverse=True)[:3]
-    top_expense_1 = f"{top_expenses[0].name} of amount ₹{top_expenses[0].amount:,.2f} on {top_expenses[0].date.strftime('%Y-%m-%d')}" if len(top_expenses) > 0 else "N/A"
-    top_expense_2 = f"{top_expenses[1].name} of amount ₹{top_expenses[1].amount:,.2f} on {top_expenses[1].date.strftime('%Y-%m-%d')}" if len(top_expenses) > 1 else "N/A"
-    top_expense_3 = f"{top_expenses[2].name} of amount ₹{top_expenses[2].amount:,.2f} on {top_expenses[2].date.strftime('%Y-%m-%d')}" if len(top_expenses) > 2 else "N/A"
-
-    with open(template_file, 'r') as file:
-        email_body = file.read()
-
-    email_body = email_body.replace('{{Month}}', report_date.strftime('%B %Y'))
-    email_body = email_body.replace('{{total_spent}}', f"₹{total_spent:,.2f}")
-    email_body = email_body.replace('{{total_saved}}', f"₹{total_saved:,.2f}")
-    email_body = email_body.replace('{{percent_change_expenses}}', f"{percent_change_expenses:.2f}%" if isinstance(percent_change_expenses, float) else "N/A")
-    email_body = email_body.replace('{{budget_used_percent}}', f"{budget_used_percent:.2f}%")
-    email_body = email_body.replace('{{overspent_categories_names_comma_separated}}', ', '.join(overspent_categories) if overspent_categories else "None")
-    email_body = email_body.replace('{{savings_goal_reached}}', "Yes" if savings_goal_reached else "No")
-    email_body = email_body.replace('{{top_expense_1}}', top_expense_1)
-    email_body = email_body.replace('{{top_expense_2}}', top_expense_2)
-    email_body = email_body.replace('{{top_expense_3}}', top_expense_3)
-    email_body = email_body.replace('{{budget_vs_actual_chart_url}}', "")
-    email_body = email_body.replace('{{intention_breakdown_pie_url}}', "")
-    email_body = email_body.replace('{{daily_spend_line_chart_url}}', "")
-
-    subject = f"TrackX - {report_date.strftime('%B %Y')} Monthly Report"
-    return {"subject": subject, "body": email_body}
-
 
 @app.post("/send-monthly-report")
 def send_monthly_report(background_tasks: BackgroundTasks, db: Session = Depends(get_db)):
@@ -867,65 +792,9 @@ def send_monthly_report(background_tasks: BackgroundTasks, db: Session = Depends
     
     return {"message": "No report was generated."}
 
-def send_email(to_email: str, subject: str, body: str):
-    from_email = "jainnaman027@gmail.com"
-    password = os.getenv("smpt_email_pass")
-    print(os.getenv("smpt_email_pass"))
-
-    msg = MIMEMultipart()
-    msg['From'] = from_email
-    msg['To'] = to_email
-    msg['Subject'] = subject
-
-    msg.attach(MIMEText(body, 'html'))
-
-    try:
-        server = smtplib.SMTP('smtp.gmail.com', 587)
-        server.starttls()
-        server.login(from_email, password)
-        server.sendmail(from_email, to_email, msg.as_string())
-        server.quit()
-        print("Email sent successfully")
-    except Exception as e:
-        print(f"Failed to send email: {e}")
-
-def scheduled_report_job():
-    """
-    Scheduled job to send monthly report if expenses were made in the previous month.
-    """
-    print("Scheduler: Running monthly report job check...")
-    db = next(get_db())
-    today = date.today()
-    first_day_of_current_month = today.replace(day=1)
-    last_day_of_previous_month = first_day_of_current_month - timedelta(days=1)
-    year = last_day_of_previous_month.year
-    month = last_day_of_previous_month.month
-
-    start_date = datetime(year, month, 1)
-    end_date = start_date + relativedelta(months=1)
-
-    expenses_count = db.query(models.Expense).filter(
-        models.Expense.date >= start_date,
-        models.Expense.date < end_date
-    ).count()
-
-    if expenses_count > 0:
-        print(f"Scheduler: Found {expenses_count} expenses for {month}/{year}. Sending report.")
-        report_data = _send_monthly_report_logic(db, year, month)
-        if report_data:
-            try:
-                send_email("jainnaman027@gmail.com", report_data["subject"], report_data["body"])
-                print("Scheduler: Email sent successfully.")
-            except Exception as e:
-                print(f"Scheduler: Failed to send email: {e}")
-    else:
-        print(f"Scheduler: No expenses found for {month}/{year}. Report not sent.")
-    db.close()
-
 scheduler = BackgroundScheduler(daemon=True)
 scheduler.add_job(scheduled_report_job, CronTrigger(day=1, hour=6, minute=0))
 scheduler.start()
-
 atexit.register(lambda: scheduler.shutdown())
 
 class SavingGoalBase(BaseModel):
