+++ plans/env_dependency.md
@@ -1,0 +1,12 @@
+## chart to image impl has env dependent impl
+## current we did for windows
+The reason for this is subtle but critical for how npm works on Windows. While on Linux and macOS, npm creates a direct executable file in the .bin directory, on Windows, it creates a set of command files instead:
+c-to-i (a shell script)
+c-to-i.cmd (a Windows command script)
+c-to-i.ps1 (a PowerShell script)
+Our Python os.path.exists() check fails because it's looking for a file named c-to-i, not a command. The shell=True argument in subprocess.run is what allows us to run these .cmd scripts as if they were executables, but we need to point to the correct file.
+The Solution
+The fix is to explicitly use the .cmd version of the executable in our Python code. This will work reliably on Windows environments.
+I will now modify service/chart_service.py to append .cmd to the executable path.
+I've adjusted the code to use the correct c-to-i.cmd script for Windows. This should finally resolve the "path not found" error.
+Please hit the API endpoint one more time. You should see DEBUG: c-to-i exists? True in the console, followed by the successful generation of the charts. Let me know the outcome
