+++ requirements.txt
@@ -13,3 +13,4 @@ python-multipart
 joblib
 matplotlib
 numpy
+quickchart.io
