+++ service/chart_service.py
@@ -1,0 +1,95 @@
+from sqlalchemy.orm import Session
+import models
+from sqlalchemy import func
+from datetime import datetime
+from dateutil.relativedelta import relativedelta
+from quickchart import QuickChart
+
+CATEGORIES = {
+    1: "Food", 2: "Housing", 3: "Transportation", 4: "Personal",
+    5: "Utility", 6: "Recreation", 7: "Health", 8: "Savings", 9: "Debt"
+}
+
+def _generate_chart_image(chart_config: dict, filename: str) -> str:
+    """
+    Generates a chart image from a Chart.js config using QuickChart.io,
+    and returns the public URL.
+    """
+    qc = QuickChart()
+    qc.width = 500
+    qc.height = 300
+    qc.config = chart_config
+
+    try:
+        url = qc.get_url()
+        print(f"Generated QuickChart URL for {filename}: {url}")
+        return url
+    except Exception as e:
+        print(f"Error generating chart with QuickChart for '{filename}': {e}")
+        return ""
+
+def generate_intention_breakdown_chart(db: Session, year: int, month: int) -> str:
+    start_date = datetime(year, month, 1)
+    end_date = start_date + relativedelta(months=1)
+    expenses = db.query(models.Expense).filter(models.Expense.date >= start_date, models.Expense.date < end_date).all()
+    
+    intention_totals = {"Need": 0, "Want": 0, "Saving": 0}
+    for expense in expenses:
+        intention_totals[expense.intention] += expense.amount
+    
+    total = sum(intention_totals.values())
+    data = [round((intention_totals[key] / total) * 100 if total > 0 else 0, 1) for key in ['Need', 'Want', 'Saving']]
+
+    chart_config = {
+        "type": 'pie',
+        "data": {
+            "labels": ['Need', 'Want', 'Saving'],
+            "datasets": [{"data": data, "backgroundColor": ['#2196F3', '#FFC107', '#9ACD32']}]
+        },
+        "options": {"plugins": {"legend": {"position": 'top'}, "title": {"display": True, "text": 'Expense Intention Breakdown (%)'}}}
+    }
+    
+    filename = f"intention_breakdown_{year}_{month}.png"
+    return _generate_chart_image(chart_config, filename)
+
+def generate_daily_spend_chart(db: Session, year: int, month: int) -> str:
+    start_date = datetime(year, month, 1)
+    end_date = start_date + relativedelta(months=1)
+    daily_expenses = db.query(models.Expense.date, func.sum(models.Expense.amount).label('total')) \
+        .filter(models.Expense.date >= start_date, models.Expense.date < end_date) \
+        .group_by(models.Expense.date).order_by(models.Expense.date).all()
+    
+    chart_config = {
+        "type": 'line',
+        "data": {
+            "labels": [day.date.strftime("%d %b") for day in daily_expenses],
+            "datasets": [{"label": 'Daily Expenses', "data": [float(day.total) for day in daily_expenses], "fill": False, "borderColor": 'rgb(75, 192, 192)', "tension": 0.1}]
+        },
+        "options": {"scales": {"y": {"beginAtZero": True}}, "plugins": {"title": {"display": True, "text": 'Daily Expense Variation'}}}
+    }
+    filename = f"daily_spend_{year}_{month}.png"
+    return _generate_chart_image(chart_config, filename)
+
+def generate_budget_vs_actual_chart(db: Session, year: int, month: int) -> str:
+    start_date = datetime(year, month, 1)
+    end_date = start_date + relativedelta(months=1)
+    
+    category_totals_q = db.query(models.Expense.category_id, func.sum(models.Expense.amount).label('total')) \
+        .filter(models.Expense.date >= start_date, models.Expense.date < end_date) \
+        .group_by(models.Expense.category_id).all()
+    category_totals = {CATEGORIES[item.category_id]: item.total for item in category_totals_q}
+    
+    labels = list(category_totals.keys())
+    actual_data = list(category_totals.values())
+    
+    chart_config = {
+        "type": 'doughnut',
+        "data": {
+            "labels": labels,
+            "datasets": [{"data": actual_data, "backgroundColor": ['#9ACD32', '#36A2EB', '#FFCE56', '#4BC0C0', '#9966FF', '#FF9F40', '#FF6384', '#4CAF50', '#FF5252']}]
+        },
+        "options": {"plugins": {"legend": {"position": 'top'}, "title": {"display": True, "text": 'Monthly Expenses by Category'}}}
+    }
+        
+    filename = f"budget_vs_actual_{year}_{month}.png"
+    return _generate_chart_image(chart_config, filename)
